requirements.txt
GitPython
requests
PyJWT
